+++ spec/integration/karma_configs/invalidRackConfig.js
@@ -11,15 +11,14 @@ module.exports = function (config) {
         // available frameworks: https://npmjs.org/browse/keyword/karma-adapter
         frameworks: ['opal'],
 
-        middleware: ['opal_sourcemap'],
-
-        opal: {
-            sourceMapResults: true
-        },
         // list of files / patterns to load in the browser
         files: [
             'spec/**/*_spec.rb'
         ],
+
+        opal: {
+            mriRequires: ['foobar']
+        },
 
         // list of files to exclude
         exclude: [],
@@ -33,7 +32,7 @@ module.exports = function (config) {
         // test results reporter to use
         // possible values: 'dots', 'progress'
         // available reporters: https://npmjs.org/browse/keyword/karma-reporter
-        reporters: ['progress', 'specjson'],
+        reporters: ['progress', 'orspec-modified-json'],
 
         specjsonReporter: {
             outputFile: 'test_run.json'
@@ -58,7 +57,7 @@ module.exports = function (config) {
 
         // start these browsers
         // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
-        browsers: ['PhantomJS'],
+        browsers: ['PhantomJS2'],
 
 
         // Continuous Integration mode