+++ spec/integration/missing_require/main_spec.rb
@@ -1,0 +1,9 @@
+require 'missing_file'
+
+describe 'something' do
+  subject { 42 }
+
+  context 'nested' do
+    it { is_expected.to eq 42 }
+  end
+end