+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+* Add in stacktrace-js for source maps in stack traces and hide Opal/Opal-RSpec/Karma code from test failure traces
+* Examples filtered out using `filter_run_including` now show up as skipped in Karma results
+* Remove undocumented source map results feature
+* Improve error handling for cases where Sprockets can't find an asset or the Rack server dies before starting up
+
 ## 1.0.10 (27 January 2016)
 
 * Do a Bundler.require by default to reduce the number of additional MRI/Rack side require statements necessary for Opal load paths to be correct